+++ lib/caylir/graph/config.ex
@@ -13,15 +13,5 @@ defmodule Caylir.Graph.Config do
     defaults
     |> Keyword.merge(Application.get_env(otp_app, graph, []))
     |> Keyword.put(:otp_app, otp_app)
-    |> validate!(graph)
   end
-
-  defp validate!([otp_app: _], graph) do
-    # single key match only possible if both
-    # - empty environment config
-    # - missing inline defaults
-    raise ArgumentError, "graph #{inspect(graph)} has no usable configuration"
-  end
-
-  defp validate!(config, _), do: config
 end