+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Enhancements
     - A missing `host` configuration will now default to `"localhost"`
+    - A missing `port` configuration will now default to `64210`
     - Queries are executed directly in the calling process without passing through a process pool
 
 ## v0.11.1 (2019-11-20)